mostly done

+++ app/Http/Livewire/TtmaLivewire.php
@@ -21,11 +21,16 @@ class TtmaLivewire extends Component
     public $ttma_id;
     public $search;
     public $duration;
+    public $message;
+    public $selected;
+    public $comment;
 
     protected $rules = [
-        'subject' => ['required', 'min:5'],
-        'user_id' => ['required'],
-        'output' => ['required', 'min:5'],
+        'subject' => ['nullable', 'required_if:selected,assign', 'min:5'],
+        'user_id' => ['nullable', 'required_if:selected,assign'],
+        'output' => ['nullable', 'required_if:selected,assign', 'min:5'],
+        'message' => ['nullable', 'required_if:selected,message', 'min:5'],
+        'comment' => ['nullable', 'required_if:selected,disapprove', 'min:5'],
     ];
 
     protected  $queryString = ['search'];
@@ -39,12 +44,27 @@ class TtmaLivewire extends Component
         $ttmas = Ttma::query();
 
         if ($search) {
-            $ttmas->whereHas('user', function (\Illuminate\Database\Eloquent\Builder $query) use ($search) {
-                return $query->where('name', 'LIKE', '%' . $search . '%');
-            })
-                ->orWhere('subject', 'LIKE', '%' . $search . '%')
-                ->orWhere('output', 'LIKE', '%' . $search . '%')
-                ->get();
+
+            if($this->duration) {
+                $ttmas->where(function ($query) use ($search) {
+                    $query->whereHas('user', function (\Illuminate\Database\Eloquent\Builder $query) use ($search) {
+                        return $query->where('name', 'LIKE', '%' . $search . '%');
+                    })
+                        ->orWhere('subject', 'LIKE', '%' . $search . '%')
+                        ->orWhere('output', 'LIKE', '%' . $search . '%');
+                })->where('duration_id', $this->duration->id)
+                ->where('head_id', Auth::user()->id)
+                ->get();;
+            } else {
+                $ttmas->where(function ($query) use ($search) {
+                    $query->whereHas('user', function (\Illuminate\Database\Eloquent\Builder $query) use ($search) {
+                        return $query->where('name', 'LIKE', '%' . $search . '%');
+                    })
+                        ->orWhere('subject', 'LIKE', '%' . $search . '%')
+                        ->orWhere('output', 'LIKE', '%' . $search . '%');
+                })->where('head_id', Auth::user()->id)
+                    ->get();
+            }
         }
 
         if($this->duration) {
@@ -60,7 +80,7 @@ class TtmaLivewire extends Component
             ]);
         }
         return view('livewire.ttma-livewire', [
-            'ttmas' => $ttmas->paginate(10),
+            'ttmas' => $ttmas->where('head_id', Auth::user()->id)->paginate(10),
             'assignments' => Ttma::orderBy('created_at', 'DESC')->paginate(10),
         ]);
     }
@@ -72,6 +92,7 @@ class TtmaLivewire extends Component
 
     public function save()
     {
+
         $this->validate();
 
         if ($this->ttma_id) {
@@ -80,7 +101,7 @@ class TtmaLivewire extends Component
             $userOld = User::where('id', $ttma->user_id)->first();
 
             foreach ($userOld->notifications as $notification) {
-                if($notification->data['ttma_id'] == $ttma->id){
+                if(isset($notification->data['ttma_id']) && $notification->data['ttma_id'] == $ttma->id){
                     $notification->delete();
                 }
             }
@@ -116,6 +137,45 @@ class TtmaLivewire extends Component
         $this->dispatchBrowserEvent('close-modal');
     }
 
+    public function disapproved() {
+        $this->validate();
+
+        $ttma = Ttma::where('id', $this->ttma_id)->first();
+    
+        $user = User::where('id', $ttma->user_id)->first();
+
+        $user->notify(new AssignmentNotification($ttma, 'Disapproved'));
+
+        Ttma::where('id', $this->ttma_id)->update([
+            'message' => null,
+            'comments' => $this->comment,
+        ]);
+
+        session()->flash('message', 'Sent Successfully!');
+        $this->resetInput();
+        $this->dispatchBrowserEvent('close-modal');
+    }
+
+    public function message() {
+
+        $this->validate();
+
+        $ttma = Ttma::where('id', $this->ttma_id)->first();
+    
+        $user = User::where('id', $ttma->head_id)->first();
+
+        $user->notify(new AssignmentNotification($ttma, 'Done'));
+
+        Ttma::where('id', $this->ttma_id)->update([
+            'message' => $this->message,
+            'comments' => null,
+        ]);
+
+        session()->flash('message', 'Sent Successfully!');
+        $this->resetInput();
+        $this->dispatchBrowserEvent('close-modal');
+    }
+
     public function done()
     {
         $ttma = Ttma::where('id', $this->ttma_id)->first();
@@ -133,8 +193,9 @@ class TtmaLivewire extends Component
         $this->dispatchBrowserEvent('close-modal');
     }
 
-    public function select($ttma_id, $category = null)
+    public function select($select, $ttma_id, $category = null)
     {
+        $this->selected = $select;
         $this->ttma_id = $ttma_id;
 
         if ($category == 'edit') {
@@ -144,6 +205,7 @@ class TtmaLivewire extends Component
             $this->subject = $data->subject;
             $this->user_id = $data->user_id;
             $this->output = $data->output;
+            $this->message = $data->message;
         }
     }
 
@@ -154,7 +216,7 @@ class TtmaLivewire extends Component
         $user = User::where('id', $ttma->user_id)->first();
 
         foreach ($user->notifications as $notification) {
-            if($notification->data['ttma_id'] == $ttma->id){
+            if(isset($notification->data['ttma_id']) && $notification->data['ttma_id'] == $ttma->id){
                 $notification->delete();
             }
         }
@@ -172,6 +234,11 @@ class TtmaLivewire extends Component
         $this->user_id = '';
         $this->output = '';
         $this->ttma_id = '';
+        $this->search = '';
+        $this->duration = '';
+        $this->message = '';
+        $this->selected = '';
+        $this->comment = '';
     }
 
     public function closeModal()
